+++ CHANGELOG.md
@@ -503,6 +503,7 @@ and try to associate the created connection with this project automatically.
 * [#1422](https://github.com/clojure-emacs/cider/issues/1422): Don't display mismatching parens error on incomplete expressions in REPL buffers.
 * [#1412](https://github.com/clojure-emacs/cider/issues/1412): nREPL messages for separate sessions are tracked in separate buffers.
 * Removed `cider-switch-to-repl-command`.
+* Renamed `cider-default-repl-command` to `cider-jack-in-default`.
 
 ### Bugs fixed